+++ src/main/scala/beam/agentsim/agents/choice/mode/ModeChoiceUniformRandom.scala
@@ -20,10 +20,5 @@ class ModeChoiceUniformRandom(val beamServices: BeamServices) extends ModeChoice
 
   override def utilityOf(alternative: EmbodiedBeamTrip): Double = 0.0
 
-  override def utilityOf(
-    mode: Modes.BeamMode,
-    cost: BigDecimal,
-    time: BigDecimal,
-    numTransfers: Int
-  ): Double = 0.0
+  override def utilityOf(mode: Modes.BeamMode, cost: Double, time: Double, numTransfers: Int): Double = 0.0
 }
